Move database initialize package to initialize package
