+++ server/src/initialize/testdatabase.go
@@ -8,7 +8,6 @@ import (
 	"github.com/testcontainers/testcontainers-go"
 	"github.com/testcontainers/testcontainers-go/modules/postgres"
 	"github.com/uptrace/bun"
-	"go.uber.org/zap/zapcore"
 )
 
 const POSTGRES_IMAGE = "postgres:17.5-alpine"
@@ -41,7 +40,7 @@ func StartTestDatabase() (*postgres.PostgresContainer, *bun.DB) {
 		log.Fatalf("Failed to initialize database %s", err)
 	}
 
-	bunDb := InitializeBun(db, zapcore.DebugLevel) // setup bun
+	bunDb := InitializeBun(db, true) // setup bun
 
 	return pgcontainer, bunDb
 }
